Move text commands to their own modules
